liblepton: Fix two compilation warnings.
Fixes warnings by gcc as follows:

  warning: format ‘%lu’ expects argument of type ‘long unsigned
  int’, but argument 3 has type ‘gsize {aka unsigned int}’
  [-Wformat=]

+++ liblepton/src/s_textbuffer.c
@@ -229,7 +229,7 @@ s_textbuffer_next_line (TextBuffer *tb)
 
     if (verbose_loading)
     {
-      fprintf (stderr, "%-4lu: %s", tb->linenum, line);
+      fprintf (stderr, "%-4lu: %s", (unsigned long) tb->linenum, line);
     }
   }
